Move memory report function to own file.
