+++ vlasiator.cpp
@@ -50,6 +50,7 @@
 #include "grid.h"
 #include "iowrite.h"
 #include "ioread.h"
+#include "memory_report.h"
 
 #include "object_wrapper.h"
 #include "fieldsolver/gridGlue.hpp"
@@ -774,7 +775,7 @@ int simulate(int argn,char* args[]) {
    
 
    phiprof::Timer reportMemTimer {"report-memory-consumption"};
-   report_node_memory_consumption(mpiGrid);
+   report_memory_consumption(mpiGrid);
    reportMemTimer.stop();
    
    unsigned int computedCells=0;
@@ -860,7 +861,7 @@ int simulate(int argn,char* args[]) {
       if (P::diagnosticInterval != 0 && P::tstep % P::diagnosticInterval == 0) {
          phiprof::Timer memTimer {"memory-report"};
          memTimer.start();
-         report_node_memory_consumption(mpiGrid);
+         report_memory_consumption(mpiGrid);
          memTimer.stop();
          phiprof::Timer cellTimer {"cell-count-report"};
          cellTimer.start();
